Re-enable shadow mapping in the hangar.
The hangar shadow map is computed using OPT geometry, which is indexed. Hence
we need to use the proper input layout or the shadow map won't render.

Further, I think (?) the error message reported by DX in this area is related
to the type used in the VertexShader input struct. It should be an uint4 instead
of an int4 -- but it should still be an integer type as it's used to index the
geometry.

+++ impl11/shaders/HangarShadowMapVS.hlsl
@@ -6,7 +6,7 @@ Buffer<float3> g_vertices : register(t0);
 
 struct VertexShaderInput
 {
-	float4 index : POSITION;
+	uint4 index : POSITION;
 };
 
 struct SHADOW_PS_INPUT
